Create project skeleton
This project skeleton is mirrored from the Truth Goggles skeleton that
we started to build out over at https://github.com/badideafactory/truthgoggles

It is intended to ultimately include the following stack:

- GraphQL + Sequelize + Postgres
- React + Parcel
- ESLint + Babel
- Express

Instead of having the server and the client be two separate projects
within the repository, they are divided code bases that are part of the
same package.  This may need to change eventually (at which point we may
actually make two separate repositories) but we'll stick with this as
long as possible.

This commit also creates some shell documentation files, which should be
fleshed out asap

Issue #1 Set up core stack