Add opened captions
Opened captions needs to be cleaned up a bit, but this at least adds it
and connects to the core server.  It also overrides the default word
handler with one of our own (which... I guess is how you do things)

Eventually we need to get the creator of opened captions to make it
easier to work with the data on the server side (cough.)  In the mean
time this will do.

We don't actually do anything with the words yet; that will come next
when we add named entity recognition.

Issue #3 Opened captions integration