Add spacy to the project
The eagle eyed will notice that spacy is not actually a node dependency.
That's because apparently node is for making rapid prototypes in React
and not for serious business.  Well FINE, we can have it BOTH WAYS!

This commit adds Spacy, which is an amazing python library for NLP.
There are other NLP tools out there, but this seems to be the best these
days.  You can learn more about it here: https://spacy.io/

Unfortunately there is now a third project within this project.  That's
just how we roll around here.  Again, maybe some day this will be broken
out into its own thing.  For now it runs.

The api shell was taken from https://github.com/explosion/spacy-services/tree/master/displacy
but it will almost certainly need to be adapted over time.

Issue #4 Set up named entity extraction