Add basic use of nlp
This comines opened captions with the spacy API in order to run TV
captions and find dependencies in real time.  It's not refined, it isn't
quite what we need for this project, but it is freakin cool.

Next steps will be to do a better job of organizing the text and tokens,
after which we'll want to tie this all into models and schemas and
databases.

Issue #4 set up named entity extraction