Add initial models and schemas
This is a mega commit, sorry about that.

We add three new types of entitiy: channel, sentence, and named entity.
For now we aren't actually using channel (This should change as Opened
Captions expands).  A sentence represents a sentence from a real time
caption stream, and a named entity represents a named entity found in
a caption stream.

This commit creates the migrations, models, and graphQL schemas for
these three entities.  It also exposes the data via a graphQL api and
uses that API for a very bare bones demo.

Issue #10 Set up storage for sentences
Issue #11 Set up graphql schema for sentence entities