Move public directory
The organization of the directories is a little weird no matter how you
slice it, but moving the public (static) assets outside of the client
source is going to make it easier to set up hooks and also might be
easier for new developers to understand what's going on.

This will be more important before we start building out react
components.