Add client dependencies
This demo uses all sorts of crazy things.  React and graphQL for one,
but also with that comes stuff like bundlers (parcel).

This commit installs a whole slew of new dependencies.

Unfortunately this commit is part of Way Too Much At Once which means
that it includes dependencies for a series of future commits ranging
from backend all the way to frontend D3.

Issue #13