Baseline react demonstration
This commit is way too large, but there isn't a huge amount of value in
breaking it up at this point... I always say that.

Anyway, this contains a very basic interface to show off the talking
point tracker functionality as it stands. It even includes a VERY
IMPRESSIVE logo which is safe to consider 100% final.

This is written in react, and leverages graphQL to load in the data
itself.

Issue #13