Add subscription skeleton
This follows the first few steps in the [Apollo
Documentation](https://www.apollographql.com/docs/graphql-subscriptions/subscriptions-to-schema.html)
for adding subscriptions to graphql.  It sets up the schema to handle
subscriptions and also creates the very initial integration with the
PubSub implementation baked into the graphql-subscriptions library.

As noted in that library, it is only appropriate for demos, not
production.  We will eventually want to replace it with a more robust
PubSub implementation.

Issue #44
Issue #43