Broadcast hovered entity
When hovering on an entity in the entity list, broadcast that entity up
to the App state store.

And on that note, go ahead and move the Dashboard state up to App. This
may not be necessary, but I’m going to hold it all in one place until
it proves necessary/profitable to do otherwise.

Also, make a few consistency syntax tweaks.

Affects #32