Add a real entity frequency graph
Divide the entity appearances into time intervals (based on the overall
interval scope) and render a graph of lines corresponding to each
entity. When an entity is activated, activate its corresponding line,
and vice versa. (Hover-highlighting of graph lines is currently
disabled.)

It also improves the layout to better accomodate the graph, espcially
at smaller viewport sizes.

This version is slow and not as interactive as I’d like, and the tick
interval calculation method is very simplistic and needs improving.

Affects #33